+++ CHANGELOG.md
@@ -7,6 +7,7 @@
 ### Notebooks
 
 - Added new notebook for `Percentiles` (`notebooks/Analyses/Percentiles.ipynb`)
+- Added new notebook for `LinearInterpolation` (`notebooks/GapFilling/LinearInterpolation.ipynb`)
 - Updated notebook for `DaytimeNighttimeFlag` (`notebooks/CalculateVariable/DaytimeNighttimeFlag.ipynb`)
 
 ## v0.71.4 | 20 Mar 2024