+++ CHANGELOG.md
@@ -4,10 +4,14 @@
 
 ## v0.71.6 | 23 Mar 2024
 
+![DIIVE](images/analysesZaggregatesInQuantileClassesOfXY_diive_v0.71.6.png)
+
 ### Notebooks
 
 - Added new notebook for `Percentiles` (`notebooks/Analyses/Percentiles.ipynb`)
 - Added new notebook for `LinearInterpolation` (`notebooks/GapFilling/LinearInterpolation.ipynb`)
+- Added new notebook for calculating z-aggregates in quantiles (classes) of x and
+  y  (`notebooks/Analyses/CalculateZaggregatesInQuantileClassesOfXY.ipynb`)
 - Updated notebook for `DaytimeNighttimeFlag` (`notebooks/CalculateVariable/DaytimeNighttimeFlag.ipynb`)
 
 ## v0.71.4 | 20 Mar 2024