+++ CHANGELOG.md
@@ -2,6 +2,18 @@
 
 ![DIIVE](images/logo_diive1_256px.png)
 
+## v0.71.6 | 23 Mar 2024
+
+![DIIVE](images/analysesZaggregatesInQuantileClassesOfXY_diive_v0.71.6.png)
+
+### Notebooks
+
+- Added new notebook for `Percentiles` (`notebooks/Analyses/Percentiles.ipynb`)
+- Added new notebook for `LinearInterpolation` (`notebooks/GapFilling/LinearInterpolation.ipynb`)
+- Added new notebook for calculating z-aggregates in quantiles (classes) of x and
+  y  (`notebooks/Analyses/CalculateZaggregatesInQuantileClassesOfXY.ipynb`)
+- Updated notebook for `DaytimeNighttimeFlag` (`notebooks/CalculateVariable/DaytimeNighttimeFlag.ipynb`)
+
 ## v0.71.5 | 22 Mar 2024
 
 ### Changes